// front/lib/generated/asset_registry/types/assetType.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum AssetType {
  Equity,
  RevenueShare,
  Royalty,
  RealEstate,
  Debt,
  Commodity,
  PhysicalGood,
  Other,
}

export type AssetTypeArgs = AssetType;

export function getAssetTypeEncoder(): FixedSizeEncoder<AssetTypeArgs> {
  return getEnumEncoder(AssetType);
}

export function getAssetTypeDecoder(): FixedSizeDecoder<AssetType> {
  return getEnumDecoder(AssetType);
}

export function getAssetTypeCodec(): FixedSizeCodec<AssetTypeArgs, AssetType> {
  return combineCodec(getAssetTypeEncoder(), getAssetTypeDecoder());
}